Update patch set 13
Patch Set 13:

auto-setup for --setup flag now runs only if an "install-builds-deps-android.sh approved" ubuntu is found. If not it returns an hint to a webpage where it's possible to find out how to install manually needed packages on other platforms. BTW this --setup it's not a big deal breaker, it can be removed if we want (@jessica ?)

Patch-set: 13