Update patch set 2
Patch Set 2:

I disagree on both accounts. The other patch is an addition to functionality on top of this ones stated purpose. Mergingt this patch doesn't override the validity of that other patch. 

As for the pull request, while I admire the sentiment, regional confits are not a limitation factor, but a sanity factor. As a user in X region, its higky unlikely I would ever care for seeing GB, RU or other regions search pages. I'd rather fine tune these as required than remove all notion of the regional configs.

Patch-set: 2