Update patch set 2
Patch Set 2:

Having long list is pretty useless because of reasons I've already explained. And of an user is so "requestful" to change his/her locale to English I think (s)he'd understand English search results.
Engines such as google and duckduck go allow you to change the "locale" of the search engine anytime you want with cookies, and I think bing and yahoo have similar options too.
This is also the config we're shipping on cos builds with cyngn browser

Patch-set: 2